Improve Space2D and Space3D algorithms
* Don't create conclusions just using the first and last location in the
direction chain. Prioritize choosing locations in the middle of the
chain, otherwise the game can be strategized by starting at one of the
conclusion's locations and working your way to the opposite end.

* Don't only use the "opposite" direction when creating a false premise.
This also leads to strategizing, where if you can prove 1 dimension is
correct, then the entire conclusion is correct.